+++ src/libs/antares/study/parts/common/series.h
@@ -46,7 +46,7 @@ public:
 
     double getAvailablePower(unsigned int hour, unsigned int year) const;
 
-    double getSeriesIndex(unsigned int hour, unsigned int year) const;
+    uint getSeriesIndex(unsigned int hour, unsigned int year) const;
 
 public:
     /*!
